Move plan generation control to control package
