+++ org.opentosca.container.api/src/main/java/org/opentosca/container/api/controller/CsarController.java
@@ -47,7 +47,7 @@ import org.opentosca.container.api.dto.CsarDTO;
 import org.opentosca.container.api.dto.CsarListDTO;
 import org.opentosca.container.api.dto.request.CsarTransformRequest;
 import org.opentosca.container.api.dto.request.CsarUploadRequest;
-import org.opentosca.container.api.service.PlanGenerationService;
+import org.opentosca.container.control.plan.PlanGenerationService;
 import org.opentosca.container.api.service.Utils;
 import org.opentosca.container.api.util.ModelUtil;
 import org.opentosca.container.connector.winery.WineryConnector;
