+++ org.opentosca.container.api/src/main/java/org/opentosca/container/api/controller/ServiceTemplateController.java
@@ -27,8 +27,8 @@ import io.swagger.annotations.ApiParam;
 import org.opentosca.container.api.dto.ServiceTemplateDTO;
 import org.opentosca.container.api.dto.ServiceTemplateListDTO;
 import org.opentosca.container.api.dto.request.ServiceTransformRequest;
-import org.opentosca.container.api.service.PlanGenerationService;
-import org.opentosca.container.api.service.PlanGenerationService.AdaptationPlanGenerationResult;
+import org.opentosca.container.control.plan.PlanGenerationService;
+import org.opentosca.container.control.plan.PlanGenerationService.AdaptationPlanGenerationResult;
 import org.opentosca.container.api.service.InstanceService;
 import org.opentosca.container.api.service.NodeTemplateService;
 import org.opentosca.container.api.service.PlanInvokerService;
