+++ org.opentosca.container.control/src/main/java/org/opentosca/container/control/plan/PlanGenerationService.java
@@ -1,4 +1,4 @@
-package org.opentosca.container.api.service;
+package org.opentosca.container.control.plan;
 
 import java.nio.file.Path;
 import java.util.Collection;
