+++ org.opentosca.container.war/src/test/java/org/opentosca/container/war/tests/MigrateMyTinyToDo2MultiMyTinyToDoIntegrationTest.java
@@ -14,7 +14,7 @@ import org.eclipse.winery.model.tosca.TServiceTemplate;
 import org.junit.After;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.opentosca.container.api.service.PlanGenerationService;
+import org.opentosca.container.control.plan.PlanGenerationService;
 import org.opentosca.container.api.service.InstanceService;
 import org.opentosca.container.api.service.PlanInvokerService;
 import org.opentosca.container.control.OpenToscaControlService;
